1. Fixing infrastructure typo. 2. Adding Failure class on top of Error. 3. Adding problem details. 4. Adding code to controller and making sure tests pass. 5. Adding language ext extension. 6 Fixing payment command signature to return failure.

+++ PaymentGateway.API/Features/Payments/CreatePaymentRequest.cs
@@ -20,8 +20,8 @@
 
         public class ExpirationDateDto
         {
-            public int Year { get; set; }
-            public int Month { get; set; }
+            public short Year { get; set; }
+            public byte Month { get; set; }
         }
     }
 }
